address @dhalperi review comments
