+++ projects/batfish/src/test/java/org/batfish/bdp/BdpDataPlanePluginTest.java
@@ -1,7 +1,6 @@
 package org.batfish.bdp;
 
 import static java.util.Collections.singletonList;
-import static org.batfish.bdp.BdpOscillationExceptionMatchers.hasMessage;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
@@ -64,6 +63,7 @@ import org.batfish.datamodel.collections.NodeInterfacePair;
 import org.batfish.datamodel.collections.RoutesByVrf;
 import org.batfish.main.Batfish;
 import org.batfish.main.BatfishTestUtils;
+import org.hamcrest.FeatureMatcher;
 import org.hamcrest.Matcher;
 import org.junit.Rule;
 import org.junit.Test;
@@ -72,6 +72,24 @@ import org.junit.rules.TemporaryFolder;
 
 /** Tests for {@link BdpDataPlanePlugin}. */
 public class BdpDataPlanePluginTest {
+
+  private static class BdpOscillationExceptionMatchers {
+    private static class HasMessage extends FeatureMatcher<BdpOscillationException, String> {
+
+      public HasMessage(Matcher<? super String> subMatcher) {
+        super(subMatcher, "message", "message");
+      }
+
+      @Override
+      protected String featureValueOf(BdpOscillationException actual) {
+        return actual.getMessage();
+      }
+    }
+
+    public static HasMessage hasMessage(Matcher<? super String> subMatcher) {
+      return new HasMessage(subMatcher);
+    }
+  }
 
   private static String TESTRIGS_PREFIX = "org/batfish/grammar/cisco/testrigs/";
 
@@ -397,11 +415,21 @@ public class BdpDataPlanePluginTest {
     settings.setBdpPrintAllIterations(true);
     settings.setBdpPrintOscillatingIterations(true);
     settings.setBdpRecordAllIterations(true);
+
+    /*
+     * Data plane computation should fail due to oscillation. Detailed information about the
+     * oscillation should appear in the thrown exception.
+     */
     _thrown.expect(
         allOf(
             ImmutableList.of(
                 isA(BdpOscillationException.class),
-                hasMessage(containsString("Changed routes (iteration 2 ==> 3)")))));
+                BdpOscillationExceptionMatchers.hasMessage(
+                    containsString("Changed routes (iteration 2 ==> 3)")))));
+    /*
+     *  Assertions in test function below are not reached. In this test we only care about a proper
+     *  exception being thrown during data plane computation.
+     */
     testBgpOscillationRecovery(settings);
   }
 
@@ -414,11 +442,22 @@ public class BdpDataPlanePluginTest {
     settings.setBdpPrintAllIterations(true);
     settings.setBdpPrintOscillatingIterations(true);
     settings.setBdpRecordAllIterations(false);
+
+    /*
+     * Data plane computation should fail due to oscillation. Despite not recording enough info for
+     * proper debugging initially, detailed information about the oscillation should still appear
+     * in the thrown exception.
+     */
     _thrown.expect(
         allOf(
             ImmutableList.of(
                 isA(BdpOscillationException.class),
-                hasMessage(containsString("Changed routes (iteration 2 ==> 3)")))));
+                BdpOscillationExceptionMatchers.hasMessage(
+                    containsString("Changed routes (iteration 2 ==> 3)")))));
+    /*
+     *  Assertions in test function below are not reached. In this test we only care about a proper
+     *  exception being thrown during data plane computation.
+     */
     testBgpOscillationRecovery(settings);
   }
 
@@ -431,11 +470,21 @@ public class BdpDataPlanePluginTest {
     settings.setBdpPrintAllIterations(false);
     settings.setBdpPrintOscillatingIterations(false);
     settings.setBdpRecordAllIterations(false);
+
+    /*
+     * Data plane computation should fail due to oscillation. Since printing is off, we should not
+     * see detailed information about the oscillation.
+     */
     _thrown.expect(
         allOf(
             ImmutableList.of(
                 isA(BdpOscillationException.class),
-                hasMessage(not(containsString("Changed routes (iteration 2 ==> 3)"))))));
+                BdpOscillationExceptionMatchers.hasMessage(
+                    not(containsString("Changed routes (iteration 2 ==> 3)"))))));
+    /*
+     *  Assertions in test function are not reached. In this test we only care about a proper
+     *  exception being thrown during data plane computation.
+     */
     testBgpOscillationRecovery(settings);
   }
 
@@ -447,6 +496,12 @@ public class BdpDataPlanePluginTest {
     settings.setBdpPrintAllIterations(false);
     settings.setBdpPrintOscillatingIterations(false);
     settings.setBdpRecordAllIterations(true);
+
+    /*
+     * Data plane computation should succeed despite oscillation, since we have enabled recovery.
+     * Assertions about proper data plane computation results are made in the below helper
+     * function.
+     */
     testBgpOscillationRecovery(settings);
   }
 
@@ -459,6 +514,13 @@ public class BdpDataPlanePluginTest {
     settings.setBdpPrintAllIterations(false);
     settings.setBdpPrintOscillatingIterations(false);
     settings.setBdpRecordAllIterations(false);
+
+    /*
+     * Data plane computation should succeed despite oscillation, since we have enabled recovery.
+     * We do not initially record enough information to perform recovery. Success of assertions
+     * contained in below helper function implies correct behavior of functionality to alter
+     * recording settings and rerun with enough information to perform recovery.
+     */
     testBgpOscillationRecovery(settings);
   }
 
@@ -478,7 +540,13 @@ public class BdpDataPlanePluginTest {
     settings.setBdpRecordAllIterations(bdpSettings.getBdpRecordAllIterations());
     BdpDataPlanePlugin dataPlanePlugin = new BdpDataPlanePlugin();
     dataPlanePlugin.initialize(batfish);
+
+    /*
+     * Data plane computation succeeds iff recovery is enabled. If disabled, an exception is thrown
+     * and should be expected by caller.
+     */
     dataPlanePlugin.computeDataPlane(false);
+
     SortedMap<String, SortedMap<String, SortedSet<AbstractRoute>>> routes =
         dataPlanePlugin.getRoutes();
     Prefix bgpPrefix = new Prefix("1.1.1.1/32");
@@ -504,6 +572,9 @@ public class BdpDataPlanePluginTest {
     boolean r2AsNextHop = r3NextHop.equals("r2");
     boolean r3AsNextHop = r2NextHop.equals("r3");
 
+    /*
+     * Data plane computation should succeed as follows if recovery is enabled.
+     */
     assertThat(r2MatchingRoutes.count(), equalTo(1L));
     assertThat(r3MatchingRoutes.count(), equalTo(1L));
     assertThat(routesWithR1AsNextHop, equalTo(1));
